requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and storage
pymongo==4.6.0

# Data validation and processing
pydantic[email]==2.5.0
numpy>=1.21.0
scipy>=1.11.0

# HTTP clients and requests
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9

# WebSocket support
websockets>=11.0
websocket-client>=1.6.0

# AI and ML integrations
groq>=0.4.1

# Telephony integration
piopiy

# Environment and configuration
python-dotenv>=1.0.0

# Additional dependencies for voice processing and TTS
# (Add these if needed for your WebSocket voice bot)
# google-cloud-texttospeech
# deepgram-sdk
# twilio
